tool-listings/legaldraft-temp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
reportlab==4.0.7
jinja2==3.1.2
python-docx==0.8.11
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
